SystemCode/requirements.txt
absl-py==2.1.0
accelerate==0.27.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.5
cssselect==1.2.0
ctransformers==0.2.27
cycler==0.12.1
dataclasses-json==0.6.4
datasets==2.17.0
dateparser==1.2.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
docstring_parser==0.16
et-xmlfile==1.1.0
eventregistry==9.1
exceptiongroup==1.2.1
executing==2.0.1
faiss-cpu==1.8.0
faiss-gpu==1.7.2
fastjsonschema==2.19.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.4
FlashRank==0.2.4
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozendict==2.4.2
frozenlist==1.4.1
fsspec==2023.10.0
fuzzywuzzy==0.18.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
google-pasta==0.2.0
googleapis-common-protos==1.63.0
GoogleNews==1.6.14
greenlet==3.0.3
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
h5py==3.11.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.4.0
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter_client==8.6.1
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
jupyterlab_widgets==3.0.10
keras==3.2.1
kiwisolver==1.4.5
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-google-genai==1.0.3
langchain-text-splitters==0.0.1
langsmith==0.1.49
Levenshtein==0.25.1
libclang==18.1.1
llama_cpp_python==0.2.67
lxml==5.2.1
lxml_html_clean==0.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
multitasking==0.0.11
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
newspaper3k==0.2.8
nltk==3.8.1
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.17.3
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.1
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.0
peewee==3.17.3
peft==0.8.2
pexpect==4.9.0
pillow==10.3.0
pip==24.0
platformdirs==4.2.0
plotly==5.21.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
pypdfium2==4.29.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-Levenshtein==0.25.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
qtconsole==5.5.1
QtPy==2.4.1
rank-bm25==0.2.2
rapidfuzz==3.8.1
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
requests-file==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
Send2Trash==1.8.3
sentence-transformers==2.7.0
sentencepiece==0.2.0
setuptools==69.5.1
sgmllib3k==1.0.0
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
streamlit==1.33.0
streamlit-feedback==0.1.3
sympy==1.12
tagui==1.50.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.4.0
tinycss2==1.2.1
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.38.0
triton==2.2.0
trl==0.7.10
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
typing-inspect==0.9.0
tyro==0.8.3
tzdata==2024.1
tzlocal==5.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
vaderSentiment==3.3.2
watchdog==4.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
wheel==0.43.0
widgetsnbextension==4.0.10
wikipedia==1.4.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
yfinance==0.2.38
